+++ src/main/java/konkuk/travelmate/domain/Matching.java
@@ -9,6 +9,7 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
 public class Matching {
 
     @Id
