+++ src/main/java/konkuk/travelmate/domain/Request.java
@@ -10,6 +10,7 @@ import static jakarta.persistence.FetchType.*;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
 public class Request {
 
     @Id
